+++ account-keeper-sdk/src/main/java/com/dwarfeng/acckeeper/sdk/bean/entity/FastJsonAccount.java
@@ -6,9 +6,11 @@ import com.dwarfeng.subgrade.sdk.bean.key.FastJsonStringIdKey;
 import com.dwarfeng.subgrade.stack.bean.Bean;
 import org.springframework.lang.NonNull;
 
+import java.util.Date;
+
 public class FastJsonAccount implements Bean {
 
-    private static final long serialVersionUID = 8385679921891379246L;
+    private static final long serialVersionUID = 9161141611753087374L;
 
     public static FastJsonAccount of(@NonNull Account account) {
         return new FastJsonAccount(
@@ -17,7 +19,8 @@ public class FastJsonAccount implements Bean {
                 account.isEnabled(),
                 account.getRemark(),
                 account.getSerialVersion(),
-                account.getDisplayName()
+                account.getDisplayName(),
+                account.getRegisteredDate()
         );
     }
 
@@ -36,18 +39,18 @@ public class FastJsonAccount implements Bean {
     @JSONField(name = "serial_version", ordinal = 5)
     private long serialVersion;
 
-    /**
-     * @since 1.3.1
-     */
     @JSONField(name = "display_name", ordinal = 6)
     private String displayName;
+
+    @JSONField(name = "registered_date", ordinal = 7)
+    private Date registeredDate;
 
     public FastJsonAccount() {
     }
 
     public FastJsonAccount(
             FastJsonStringIdKey key, String password, boolean enabled, String remark, long serialVersion,
-            String displayName
+            String displayName, Date registeredDate
     ) {
         this.key = key;
         this.password = password;
@@ -55,6 +58,7 @@ public class FastJsonAccount implements Bean {
         this.remark = remark;
         this.serialVersion = serialVersion;
         this.displayName = displayName;
+        this.registeredDate = registeredDate;
     }
 
     public FastJsonStringIdKey getKey() {
@@ -105,6 +109,14 @@ public class FastJsonAccount implements Bean {
         this.displayName = displayName;
     }
 
+    public Date getRegisteredDate() {
+        return registeredDate;
+    }
+
+    public void setRegisteredDate(Date registeredDate) {
+        this.registeredDate = registeredDate;
+    }
+
     @Override
     public String toString() {
         return "FastJsonAccount{" +
@@ -114,6 +126,7 @@ public class FastJsonAccount implements Bean {
                 ", remark='" + remark + '\'' +
                 ", serialVersion=" + serialVersion +
                 ", displayName='" + displayName + '\'' +
+                ", registeredDate=" + registeredDate +
                 '}';
     }
 }