+++ account-keeper-stack/src/main/java/com/dwarfeng/acckeeper/stack/bean/entity/Account.java
@@ -3,6 +3,8 @@ package com.dwarfeng.acckeeper.stack.bean.entity;
 import com.dwarfeng.subgrade.stack.bean.entity.Entity;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 
+import java.util.Date;
+
 /**
  * 账户。
  *
@@ -11,23 +13,30 @@ import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
  */
 public class Account implements Entity<StringIdKey> {
 
-    private static final long serialVersionUID = -7041696490265336859L;
+    private static final long serialVersionUID = 4480763018847028201L;
 
     private StringIdKey key;
     private String password;
     private boolean enabled;
     private String remark;
     private long serialVersion;
+
     /**
      * @since 1.3.1
      */
     private String displayName;
 
+    /**
+     * @since 1.5.1
+     */
+    private Date registeredDate;
+
     public Account() {
     }
 
     public Account(
-            StringIdKey key, String password, boolean enabled, String remark, long serialVersion, String displayName
+            StringIdKey key, String password, boolean enabled, String remark, long serialVersion, String displayName,
+            Date registeredDate
     ) {
         this.key = key;
         this.password = password;
@@ -35,6 +44,7 @@ public class Account implements Entity<StringIdKey> {
         this.remark = remark;
         this.serialVersion = serialVersion;
         this.displayName = displayName;
+        this.registeredDate = registeredDate;
     }
 
     @Override
@@ -87,6 +97,14 @@ public class Account implements Entity<StringIdKey> {
         this.displayName = displayName;
     }
 
+    public Date getRegisteredDate() {
+        return registeredDate;
+    }
+
+    public void setRegisteredDate(Date registeredDate) {
+        this.registeredDate = registeredDate;
+    }
+
     @Override
     public String toString() {
         return "Account{" +
@@ -96,6 +114,7 @@ public class Account implements Entity<StringIdKey> {
                 ", remark='" + remark + '\'' +
                 ", serialVersion=" + serialVersion +
                 ", displayName='" + displayName + '\'' +
+                ", registeredDate=" + registeredDate +
                 '}';
     }
 }